refactor: Move Time extension methods to their own files.
These are just convenience methods and not used directly by Time::Iterator.
They may go away entirely.
